animated_morph_target_scene.cpp is headerless
